requirements.txt
flask
pytest
pyyaml
requests
requests_oauthlib
